+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - 2025-MM-DD
 
+### Added
+
+- CLI: replace package name in requirements ([#372](https://github.com/Lightning-AI/utilities/pull/372))
+
+
 ### Changed
 
 -